Moves test class, 'CommandJobTest', from 'test_job.py' to 'test_command_job.py'
